Create Player.java
An example bot for the Risk, the Warlight Way KotH. This bot has simplistic behavior; it deploys all the armies given to it to a single territory and moves all the armies from a random territory to one that is adjacent to it. There may be some ideas used in this code that you can use for your convenience and edit to your liking (such as the Territory class). Here's a tip: it may be helpful to find the territories adjacent to that represented by a Territory object.